requirements.txt
# Core dependencies
pdfplumber==0.10.3
python-docx==1.1.0
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.24.3

# Web framework
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-JWT-Extended==4.5.3

# Data processing and analysis
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter notebook support
jupyter==1.0.0
ipykernel==6.25.0

# Additional NLP libraries
spacy==3.6.1
textblob==0.17.1

# Development and testing
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Model persistence
joblib==1.3.2
pickle-mixin==1.0.2

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7

# Authentication & Security
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0

# Payment processing
stripe==6.7.0

# PDF generation
reportlab==4.0.4
weasyprint==59.0

# Email
sendgrid==6.10.0

# Caching
redis==4.6.0
Flask-Caching==2.1.0

# API documentation
flask-restx==1.2.0
